debugPage→loginPageへの遷移を実装
